requirements.txt
# Dependencias para el proyecto NutriLife AI + Web3

# Machine Learning / Deep Learning
tensorflow>=2.10.0
numpy>=1.23.0
scikit-learn>=1.1.0

# Procesamiento de Imágenes
Pillow>=9.0.0
opencv-python>=4.6.0
albumentations>=1.2.0

# Visualización
matplotlib>=3.5.0
seaborn>=0.12.0

# Backend (para fases posteriores)
flask>=2.2.0
flask-cors>=3.0.10

# Web3 (para fases posteriores)
web3>=6.0.0
ipfshttpclient>=0.7.0  # 0.8.0 no está disponible aún, usar 0.7.0

# Utilidades
python-dotenv>=0.20.0
requests>=2.28.0  # Para testing de endpoints

